Moved echo protocol header into detail namespace.
