+++ include/telnetpp/options/echo/detail/echo.hpp
@@ -2,8 +2,8 @@
 
 #include "telnetpp/core.hpp"
 
-namespace telnetpp { namespace options { namespace echo {
+namespace telnetpp { namespace options { namespace echo { namespace detail {
 
 static constexpr telnetpp::u8 option = 1;
 
-}}}
+}}}}
