+++ include/telnetpp/options/echo/server.hpp
@@ -1,12 +1,12 @@
 #pragma once
 
 #include "telnetpp/options/subnegotiationless_server.hpp"
-#include "telnetpp/options/echo.hpp"
+#include "telnetpp/options/echo/detail/echo.hpp"
 
 namespace telnetpp { namespace options { namespace echo {
 
 using server = telnetpp::options::subnegotiationless_server<
-    telnetpp::options::echo::option
+    telnetpp::options::echo::detail::option
 >;
 
 }}}
